feat: include source frame in missing-persist-config dev warning
Anonymous forms have synthetic `__cx:anon:*` keys, so when the
"register({ persist: true }) without persist: configured" warning
fires the dev can't grep for the offending call. Append the
captured user-code source frame (path:line:col) inline in the
message so the warning reads like "...used on form
'__cx:anon:v-5' at (pages/foo.vue:18:18), but no `persist:`
option..." giving Chrome's console auto-linker a clickable jump.

Extracted `captureUserCallSite` + `shortenSourceFrame` from
use-abstract-form.ts (where they backed the duplicate-ambient-
provide warning) into a shared `core/dev-stack-trace.ts`. Same
regex / prefix-stripping behaviour, now reusable from any dev
warning that wants to point the reader at their own code.
